move functionality to array_context.py
